+++ assets/js/web-components/prpl-guage-progress-controller.js
@@ -173,7 +173,7 @@ class PrplGaugeProgressController {
 	}
 
 	get gaugeValue() {
-		return parseFloat( this.gauge.getAttribute( 'data-value' ) ) || 0;
+		return parseInt( this.gauge.getAttribute( 'data-value' ) ) || 0;
 	}
 	set gaugeValue( v ) {
 		v = Math.max( 0, Math.min( v, this.gaugeMax ) );
@@ -181,20 +181,20 @@ class PrplGaugeProgressController {
 	}
 
 	get gaugeMax() {
-		return parseFloat( this.gauge.getAttribute( 'data-max' ) ) || 10;
+		return parseInt( this.gauge.getAttribute( 'data-max' ) ) || 10;
 	}
 
 	_barValue( bar ) {
-		return parseFloat( bar.getAttribute( 'data-points' ) ) || 0;
+		return parseInt( bar.getAttribute( 'data-points' ) ) || 0;
 	}
 	_setBarValue( bar, v ) {
-		const max = parseFloat( bar.getAttribute( 'data-max-points' ) ) || 10;
+		const max = parseInt( bar.getAttribute( 'data-max-points' ) ) || 10;
 		v = Math.max( 0, Math.min( v, max ) );
 		bar.setAttribute( 'data-points', v );
 	}
 
 	_barMaxPoints( bar ) {
-		return parseFloat( bar.getAttribute( 'data-max-points' ) ) || 10;
+		return parseInt( bar.getAttribute( 'data-max-points' ) ) || 10;
 	}
 
 	increase( amount = 1 ) {
@@ -210,7 +210,7 @@ class PrplGaugeProgressController {
 		for ( const bar of this.progressBars ) {
 			if ( remaining <= 0 ) break;
 			const barSpace =
-				parseFloat( bar.getAttribute( 'data-max-points' ) ) -
+				parseInt( bar.getAttribute( 'data-max-points' ) ) -
 				this._barValue( bar );
 
 			const toBar = Math.min( remaining, barSpace );
